requirements.txt
# Core Flask dependencies
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
Werkzeug>=2.3.0
flask

# AI and LangChain - simplified versions
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.1

# Database
SQLAlchemy>=2.0.0

# Email and environment
python-dotenv>=0.19.0

# Document processing
docx
python-docx
PyPDF2>=3.0.0
reportlab>=4.0.0
markdown>=3.4.0

# Web scraping
requests>=2.28.0
beautifulsoup4>=4.11.0

# Core Flask Framework
Flask==2.3.0
Werkzeug==2.3.0

# Database
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.0

# User Authentication
Flask-Login==0.6.3

# Environment Variables
python-dotenv==1.0.0

# AI Integration
google-generativeai==0.3.0

# Document Processing
python-docx==0.8.11
PyPDF2==3.0.1
reportlab==4.0.4
markdown==3.5.1

# Web Scraping (for opportunities)
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Email Functionality
secure-smtplib==0.1.1

# Date and Time
python-dateutil==2.8.2

# File Handling
Pillow==10.0.1

# Security
cryptography==41.0.7

# HTTP Client
urllib3==2.0.7

# JSON Web Tokens (for future API authentication)
PyJWT==2.8.0

# Development Dependencies (optional)
# Uncomment for development
# flask-debugtoolbar==0.13.1
# pytest==7.4.3
# pytest-flask==1.2.0

# Production Dependencies (optional)
# Uncomment for production deployment
gunicorn==21.2.0
psycopg2-binary==2.9.9  # For PostgreSQL
redis==5.0.1  # For caching

# LangChain (optional - for advanced AI features)
langchain==0.0.330
langchain-google-genai==0.0.6

# Additional utilities
six==1.16.0
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
